[REXA-23] review MR
